backend/requirements.txt
# whatsapp
python-dateutil >= 2.6.0
selenium >= 3.4.3
python-axolotl
cryptography
aiohttp
six
typing
numpy
python-magic
# django
django
djangorestframework
django-cors-headers
djoser